Move EuiContext to i18n section of docs site (#3116)
