+++ src-docs/src/routes.js
@@ -67,8 +67,6 @@ import { ColorPickerExample } from './views/color_picker/color_picker_example';
 
 import { ComboBoxExample } from './views/combo_box/combo_box_example';
 
-import { ContextExample } from './views/context/context_example';
-
 import { ContextMenuExample } from './views/context_menu/context_menu_example';
 
 import { ControlBarExample } from './views/control_bar/control_bar_example';
@@ -406,7 +404,6 @@ const navigation = [
     items: [
       AccessibilityExample,
       ColorPaletteExample,
-      ContextExample,
       CopyExample,
       UtilityClassesExample,
       DelayHideExample,
