+++ src-docs/src/views/i18n/props.tsx
@@ -1,0 +1,4 @@
+import React, { FunctionComponent } from 'react';
+import { I18nShape } from '../../../../src/components/context/context';
+
+export const I18nShapeProps: FunctionComponent<I18nShape> = () => <div />;
